+++ tests/test_rpc.py
@@ -95,7 +95,6 @@ async def test_stdio_simple_args(ic):
 REASON_SOCKET = "sockets hang in 3.11"
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_simple_args(sc):
     assert await sc.call.echo("hello") == "socket: hello"
@@ -111,7 +110,6 @@ async def test_stdio_simple_kwargs(ic):
     assert await ic.call.echo(msg="hello") == "stdio: hello"
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_simple_kwargs(sc):
     assert await sc.call.echo(msg="hello") == "socket: hello"
@@ -140,7 +138,6 @@ async def test_stdio_lcg_kwargs(ic, args, kwargs, result):
     assert await ic.call.lcg(*args, **kwargs) == result
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 @pytest.mark.parametrize(("args", "kwargs", "result"), LCG_CASES)
 async def test_socket_lcg_kwargs(sc, args, kwargs, result):
@@ -159,7 +156,6 @@ async def test_stdio_seq(ic):
     assert await ic.call.echo("world") == "stdio: world"
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_seq(sc):
     assert await sc.call.echo("hello", 0.1) == "socket: hello"
@@ -178,7 +174,6 @@ async def test_stdio_par1(ic):
     assert await asyncio.gather(ic.call.echo("hello", 0.1), ic.call.echo("world")) == expected
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_par1(sc):
     expected = ["socket: hello", "socket: world"]
@@ -197,14 +192,12 @@ async def test_stdio_par2(ic):
     assert await asyncio.gather(ic.call.echo("hello"), ic.call.echo("world", 0.1)) == expected
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_par2(sc):
     expected = ["socket: hello", "socket: world"]
     assert await asyncio.gather(sc.call.echo("hello"), sc.call.echo("world", 0.1)) == expected
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 @pytest.mark.asyncio()
 async def test_socket_multi_clients(socket_server_path):
     r1, w1 = await asyncio.open_unix_connection(socket_server_path)
@@ -220,7 +213,6 @@ async def test_socket_multi_clients(socket_server_path):
             assert await asyncio.gather(c2.call.echo("h"), c1.call.echo("w", 0.1)) == expected
 
 
-@pytest.mark.skipif(sys.version_info[:2] == (3, 11), reason=REASON_SOCKET)
 def test_sync_socket_rpc_client(socket_server_path):
     with SocketSyncRPCClient(socket_server_path) as client:
         assert client.call.echo("hello", _rpc_timeout=5) == "socket: hello"